+++ src/natfeats.cpp
@@ -39,13 +39,13 @@ uint32 nf_get_id(memptr stack)
 	return 0;		/* ID with given name not found */
 }
 
-uint32 nf_call(memptr stack, bool inSuper)
+int32 nf_call(memptr stack, bool inSuper)
 {
 	uint32 fncode = ReadInt32(stack);
 	unsigned int idx = MASTERID2IDX(fncode);
 	if (idx >= nf_objs_cnt) {
 		D(bug("nf_call: wrong ID %d", idx));
-		return (uint32)-32L;	/* EINVFN - invalid function */
+		return -32L;	/* EINVFN - invalid function */
 	}
 
 	fncode = MASKOUTMASTERID(fncode);
